Add a verbose parameter to CuckooSearch
Former-commit-id: 73f425b81f91f30eef1696d69645a90ba2a49b04

+++ org.getalp.lexsema-wsd/org.getalp.lexsema-wsd-core/src/main/java/org/getalp/lexsema/wsd/experiments/cuckoo/parameters/CuckooParameters.java
@@ -50,7 +50,7 @@ public class CuckooParameters implements CuckooSolution
     public String toString()
     {
         return "levy scale = " + levyScale.currentValue + 
-                ", nests number = " + nestsNumber.currentValue + 
-                ", destroyed nests = " + destroyedNests.currentValue * nestsNumber.currentValue;
+                ", nests number = " + (int) nestsNumber.currentValue + 
+                ", destroyed nests = " + (int) (destroyedNests.currentValue * nestsNumber.currentValue);
     }
 }
